Made submodules into subclasses.
