+++ src/cv2-stubs/classes.pyi
@@ -4074,7 +4074,7 @@ cuda_Event = cuda.Event
 cuda_GpuData = cuda.GpuData
 cuda_GpuMat = cuda.GpuMat
 cuda_GpuMatND = cuda.GpuMatND
-cuda_GpuMat_Allocator = cuda.GpuMat.Allocator  # pyright: ignore   TODO
+cuda_GpuMat_Allocator = cuda.GpuMat.Allocator
 cuda_HostMem = cuda.HostMem
 cuda_SURF_CUDA = cuda.SURF_CUDA
 cuda_Stream = cuda.Stream
@@ -4297,7 +4297,7 @@ text_TextDetector = text.TextDetector
 text_TextDetectorCNN = text.TextDetectorCNN
 utils_ClassWithKeywordProperties = utils.ClassWithKeywordProperties
 utils_nested_ExportClassName = utils.nested.ExportClassName
-utils_nested_ExportClassName_Params = utils.nested.ExportClassName.Params  # pyright: ignore  TODO
+utils_nested_ExportClassName_Params = utils.nested.ExportClassName.Params
 wechat_qrcode_WeChatQRCode = wechat.WeChatQRCode
 xfeatures2d_AffineFeature2D = xfeatures2d.AffineFeature2D
 xfeatures2d_BEBLID = xfeatures2d.BEBLID
